change the folder structure.
