+++ .env
@@ -1,4 +1,3 @@
 APP_NAME=My Neshama
 
-EXPO_PUBLIC_API_URL=https://admin.ipowatch.in/api/v1
-EXPO_PUBLIC_IMAGE_URL=https://admin.ipowatch.in/storage/app/public/
+EXPO_PUBLIC_API_URL=https://admin.ipowatch.in/api/v2
