+++ api/sme.ts
@@ -26,18 +26,3 @@ export const SME_IPO_LISTED_API = async (): Promise<APIResponse> => {
         return {status: false, message: err.message}
     })
 }
-
-export const SINGLE_SME_IPO_API = async(id): Promise<APIResponse> => {
-    console.log("id=========>",`/sme_ipo/${id}`);
-    return await API.get(`/sme_ipo/${296}`).then(response => {
-        console.log("434343434343",response)
-        if(response) {
-            console.log("respons+++++++++++++++++++e",response)
-            return {status: true, message: response, data: response};
-        } else
-            console.log("84684654552138552588522525585224558+++++++++++++++++++e",response)
-            return {status: false, message: response.message};
-    }).catch(err => {
-        return {status: false, message: err.message}
-    })
-}
