+++ app/(mainline)/upcomingIpo.tsx
@@ -1,28 +1,26 @@
 import * as React from "react";
 import {useEffect, useState} from "react";
+import {MAINLINE_ALL_API} from "@/api/mainline";
 import {ThemedView} from "@/components/ThemedView";
 import {ScrollView} from "react-native-gesture-handler";
 import {Card, Divider, TouchableRipple} from "react-native-paper";
 import {ThemedText} from "@/components/ThemedText";
 import {Image, Share, StyleSheet} from "react-native";
 import {Colors} from "@/constants/Colors";
-import {SmeIpoData} from "@/types/smeipo.interface";
-import {SME_IPO_LISTED_API} from "@/api/sme";
+import {MainlineData} from "@/types/mainline.interface";
 import {BANNER_API} from "@/api/banner";
 import BannerImage from "@/components/BannerImage";
-import {baseImageURL} from "@/helper/other/url-helper";
-import {useRouter} from 'expo-router';
 import ThemedButton from "@/components/ThemedButton";
+import {router} from "expo-router";
 
-export const ListedIpo: React.FC = () => {
-    const [listedData, setListedData] = useState<SmeIpoData[]>([]);
-    const [bannerData, setBannerData] = useState<SmeIpoData[]>([]);
+export const UpcomingIpo: React.FC = () => {
+    const [upcomingData, setUpcomingData] = useState<MainlineData[]>([]);
+    const [bannerData, setBannerData] = useState<MainlineData[]>([]);
 
-    const router = useRouter();
-    const fetchListedData = async () => {
+    const fetchUpcomingData = async () => {
         try {
-            const response = await SME_IPO_LISTED_API();
-            setListedData(response.data);
+            const response = await MAINLINE_ALL_API();
+            setUpcomingData(response.data);
         } catch (error) {
             console.error("Error fetching data", error);
         }
@@ -38,11 +36,11 @@ export const ListedIpo: React.FC = () => {
     };
 
     useEffect(() => {
-        fetchListedData();
+        fetchUpcomingData();
         fetchBannerData();
     }, []);
 
-    const handleShare = async (item: SmeIpoData) => {
+    const handleShare = async (item: MainlineData) => {
         try {
             const options = {
                 message: `IPO Detail\n\nCompany Name: ${item.title}\nIPO Offer Date: ${item.open_date} to ${item.close_date}\nOffer Price: ${item.offer_price}\nLotsize: ${item.lot_size} \nIPO GMP: ${item.gmp}\n\nHey I'm using IPO Watch App to get details of IPOs.\n\nDownload Now for FREE.\n\nAndroid:\nhttps://play.google.com/store/apps/details?id=com.watch.ipo_watch`
@@ -53,24 +51,24 @@ export const ListedIpo: React.FC = () => {
         }
     };
 
-    const handleSingleOffer = (id: number) => {
-        router.push(`/singleOffer?id=${id}`);
-        console.log("id", id)
+    const handleSingleOffer = (item) => {
+        router.push(`/singleOffer?item=${(item)}`);
     };
+
     return (
         <ThemedView style={styles.mainContainer}>
             <ScrollView showsVerticalScrollIndicator={false}>
                 <BannerImage bannerData={bannerData} />
-                {listedData?.map((item, index) => (
+                {upcomingData?.map((item, index) => (
                     <Card key={index} style={styles.card}>
-                        <TouchableRipple onPress={() => handleSingleOffer(item.id)}>
+                        <TouchableRipple onPress={() => handleSingleOffer(item)}>
                             <>
                                 <ThemedView style={styles.mainBoardContainer}>
                                     <ThemedText style={styles.mainBoard}>{item.label}</ThemedText>
                                 </ThemedView>
                                 <ThemedView style={styles.header}>
                                     <ThemedView style={styles.imgContainer}>
-                                        <Image source={{uri: `${baseImageURL}/smeipo_images/${item.image}`}} style={styles.img} />
+                                        <Image source={{uri: item.image}} style={styles.img} />
                                     </ThemedView>
                                     <ThemedView style={styles.headerText}>
                                         <ThemedText type={'title'}>{item.title}</ThemedText>
@@ -108,6 +106,7 @@ export const ListedIpo: React.FC = () => {
             </ScrollView>
         </ThemedView>
     )
+
 }
 
 const styles = StyleSheet.create({
