+++ app/(smeipo)/upcomingIpo.tsx
@@ -2,7 +2,7 @@ import * as React from "react";
 import {useEffect, useState} from "react";
 import {ThemedView} from "@/components/ThemedView";
 import {ScrollView} from "react-native-gesture-handler";
-import {Card, Divider} from "react-native-paper";
+import {Card, Divider, TouchableRipple} from "react-native-paper";
 import {ThemedText} from "@/components/ThemedText";
 import {Image, Share, StyleSheet} from "react-native";
 import {Colors} from "@/constants/Colors";
@@ -10,8 +10,8 @@ import {SME_IPO_UPCOMING_API} from "@/api/sme";
 import {SmeIpoData} from "@/types/smeipo.interface";
 import {BANNER_API} from "@/api/banner";
 import BannerImage from "@/components/BannerImage";
-import {baseImageURL} from "@/helper/other/url-helper";
 import ThemedButton from "@/components/ThemedButton";
+import {router} from "expo-router";
 
 export const UpcomingIpo: React.FC = () => {
     const [upcomingData, setUpcomingData] = useState<SmeIpoData[]>([]);
@@ -50,48 +50,56 @@ export const UpcomingIpo: React.FC = () => {
         }
     };
 
+    const handleSingleOffer = (item) => {
+        router.push(`/singleOffer?item=${JSON.stringify(item)}`);
+    };
+
     return (
         <ThemedView style={styles.mainContainer}>
             <ScrollView showsVerticalScrollIndicator={false}>
                 <BannerImage bannerData={bannerData} />
                 {upcomingData?.map((item, index) => (
                     <Card key={index} style={styles.card}>
-                        <ThemedView style={styles.mainBoardContainer}>
-                            <ThemedText style={styles.mainBoard}>{item.label}</ThemedText>
-                        </ThemedView>
-                        <ThemedView style={styles.header}>
-                            <ThemedView style={styles.imgContainer}>
-                                <Image source={{uri: `${baseImageURL}/smeipo_images/${item.image}`}} style={styles.img} />
-                            </ThemedView>
-                            <ThemedView style={styles.headerText}>
-                                <ThemedText type={'title'}>{item.title}</ThemedText>
-                                <ThemedText type={'subtitle'}>Offer Date : {item.open_date} to {item.close_date}</ThemedText>
-                            </ThemedView>
-                        </ThemedView>
-                        <ThemedView style={styles.itemContainer}>
-                            <ThemedView style={styles.item}>
-                                <ThemedText>IPO PRICE</ThemedText>
-                                <ThemedText type={'subtitle'}>{item.offer_price}</ThemedText>
-                            </ThemedView>
-                            <Divider style={styles.verticalDivider} />
-                            <ThemedView style={styles.item}>
-                                <ThemedText>LOT SIZE</ThemedText>
-                                <ThemedText type={'subtitle'}>{item.lot_size}</ThemedText>
-                            </ThemedView>
-                            <Divider style={styles.verticalDivider} />
-                            <ThemedView style={styles.item}>
-                                <ThemedText>SUBSCRIBE</ThemedText>
-                                <ThemedText type={'subtitle'}>{item.subscription}</ThemedText>
-                            </ThemedView>
-                        </ThemedView>
-                        <ThemedView style={styles.itemContainer}>
-                            <ThemedView style={styles.headerText}>
-                                <ThemedText type={'subtitle'}>Exp. Premium / GMP : {item.gmp}</ThemedText>
-                            </ThemedView>
-                            <ThemedView style={styles.shareButtonContainer}>
-                                <ThemedButton onPress={() => handleShare(item)} title="Share" icon="share-variant" />
-                            </ThemedView>
-                        </ThemedView>
+                        <TouchableRipple onPress={() => handleSingleOffer(item)}>
+                            <>
+                                <ThemedView style={styles.mainBoardContainer}>
+                                    <ThemedText style={styles.mainBoard}>{item.label}</ThemedText>
+                                </ThemedView>
+                                <ThemedView style={styles.header}>
+                                    <ThemedView style={styles.imgContainer}>
+                                        <Image source={{uri: item.image}} style={styles.img} />
+                                    </ThemedView>
+                                    <ThemedView style={styles.headerText}>
+                                        <ThemedText type={'title'}>{item.title}</ThemedText>
+                                        <ThemedText type={'subtitle'}>Offer Date : {item.open_date} to {item.close_date}</ThemedText>
+                                    </ThemedView>
+                                </ThemedView>
+                                <ThemedView style={styles.itemContainer}>
+                                    <ThemedView style={styles.item}>
+                                        <ThemedText>IPO PRICE</ThemedText>
+                                        <ThemedText type={'subtitle'}>{item.offer_price}</ThemedText>
+                                    </ThemedView>
+                                    <Divider style={styles.verticalDivider} />
+                                    <ThemedView style={styles.item}>
+                                        <ThemedText>LOT SIZE</ThemedText>
+                                        <ThemedText type={'subtitle'}>{item.lot_size}</ThemedText>
+                                    </ThemedView>
+                                    <Divider style={styles.verticalDivider} />
+                                    <ThemedView style={styles.item}>
+                                        <ThemedText>SUBSCRIBE</ThemedText>
+                                        <ThemedText type={'subtitle'}>{item.subscription}</ThemedText>
+                                    </ThemedView>
+                                </ThemedView>
+                                <ThemedView style={styles.itemContainer}>
+                                    <ThemedView style={styles.headerText}>
+                                        <ThemedText type={'subtitle'}>Exp. Premium / GMP : {item.gmp}</ThemedText>
+                                    </ThemedView>
+                                    <ThemedView style={styles.shareButtonContainer}>
+                                        <ThemedButton onPress={() => handleShare(item)} title="Share" icon="share-variant" />
+                                    </ThemedView>
+                                </ThemedView>
+                            </>
+                        </TouchableRipple>
                     </Card>
                 ))}
             </ScrollView>
