+++ helper/other/url-helper.js
@@ -1,3 +1,1 @@
 export const baseURL = process.env.EXPO_PUBLIC_API_URL;
-
-export const baseImageURL = process.env.EXPO_PUBLIC_IMAGE_URL;
